client/pyproject.toml
[project]
name = "super_client"
version = "9.0.0"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "A small example package"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "grpcio-tools",
]